requirements.txt
patchright>=1.57.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
google-auth>=2.23.0
requests>=2.31.0
pyyaml>=6.0.1

